feat(manga): start add CLI
will use for add new manga in database
