+++ apps/backend/services/manga/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "manga"
-version = "0.1.1b2"
+version = "0.2.0"
 description = "Main manga reader service"
 authors = [{ name = "edpyt", email = "edrsa1321s@gmail.com" }]
 requires-python = ">=3.13"
@@ -19,8 +19,14 @@ dependencies = [
 
 [project.scripts]
 manga-api = "manga.presentation.api:start_app"
+manga-cli = "manga.presentation.cli:app"
 
 [dependency-groups]
+cli = [
+    "psycopg[binary,pool]>=3.2.10",
+    "pydantic-settings>=2.10.1",
+    "typer>=0.16.1",
+]
 dev = [
     "httpx>=0.28.1",
     "pytest>=8.4.1",
