+++ apps/backend/services/manga/tests/integration/test_cli/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+from typer.testing import CliRunner
+
+
+@pytest.fixture(scope="module")
+def runner() -> CliRunner:
+    return CliRunner()
